+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
-- portscan: Report on option descriptions that match the default descriptions
+- portscan: Report on option descriptions that fuzzy match the default descriptions
   in `Mk/bsd.options.desc.mk`.  Enabled by default but is also
   selectable with `-o option-default-descriptions`.